requirements.txt
fastapi==0.115.6
python-dotenv==1.1.1
uvicorn==0.34.0
openai==1.101.0
pymongo==4.9.2
motor==3.6.0
transformers==4.55.2
sentencepiece==0.2.1
accelerate==1.10.0
protobuf==5.29.5
sentence-transformers==5.1.0
einops==0.8.1
qdrant-client==1.13.*
unstructured[pdf]==0.18.15
streamlit==1.50.0
websocket-client==1.8.0
watchfiles==1.1.0
